size of

+++ 0x00-hello_world/6-size.c
@@ -1,15 +1,16 @@
 #include <stdio.h>
 
-/**Explain
+/**
+ * main - Entry point
  *
- *
+ * Return: Always 0 (Success)
  */
 int main(void)
 {
 	printf("Size of a char: %lu byte(s)\n", sizeof(char));
 	printf("Size of an int: %lu byte(s),\n", sizeof(int));
-	printf("Size of a long int: %lu byte(s)\n", sizeof(short));
-	printf("Size of a long long int: %u byte(s)\n", sizeof(long));
+	printf("Size of a long int: %lu byte(s)\n", sizeof(long));
+	printf("Size of a long long int: %u byte(s)\n", sizeof(double));
 	printf("Size of float: %u byte(s)\n", sizeof(float)); 
 	return (0);
 }
